requirements.txt
asgiref==3.9.1
Django==5.2.6
python-dotenv==1.0.0
sqlparse==0.5.3
tzdata==2025.2
Pillow==10.4.0
django-environ==0.12.0